requirements.txt
requests==2.28.1
python-dotenv
fastapi
uvicorn[standard]